More tests for dataset callbacks
