+++ tests/callbacks/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the frontend.callbacks package."""
